cleanup feedback suites
