+++ packages/validate/test/FeedbackMixin.test.js
@@ -1,0 +1,3 @@
+import { runFeedbackMixinSuite } from '../test-suites/FeedbackMixin.suite.js';
+
+runFeedbackMixinSuite();
